Correctly place server build code
